frontend/requirements.txt
Flask
requests
WeasyPrint
Markdown2
python-dotenv
gunicorn